Added drawer and heading section to dashboard page
