+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 import { CssBaseline } from "@mui/material";
 import { ThemeProvider } from "./components/theme-provider/ThemeProvider";
-import { DashboardPage } from "./components/pages/DashboardPage";
+import { DashboardPage } from "./components/pages/dashboard/DashboardPage";
 
 export const App = () => {
   return (
